# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "后台管理系统 API",
        "title": "Admin Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/files": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "文件列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件夹ID",
                        "name": "folderId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "文件名关键词",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.PageRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/files/upload": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "普通文件上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "文件",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文件夹ID",
                        "name": "folderId",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FileRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/files/upload/chunk": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "分片上传",
                "parameters": [
                    {
                        "type": "file",
                        "description": "分片数据",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "上传会话ID",
                        "name": "fileId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "原始文件名",
                        "name": "fileName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "分片索引",
                        "name": "chunkIndex",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "总分片数",
                        "name": "totalChunks",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "文件夹ID",
                        "name": "folderId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "MIME类型",
                        "name": "mimeType",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MsgRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/files/upload/chunk/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "分片上传状态",
                "parameters": [
                    {
                        "type": "string",
                        "description": "上传会话ID",
                        "name": "fileId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "总分片数",
                        "name": "totalChunks",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ChunkStatusRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/files/upload/merge": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "合并分片",
                "parameters": [
                    {
                        "description": "合并参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin-server_internal_model.MergeFileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FileRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/files/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "文件详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.FileRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "删除文件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.MsgRespWrapper"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/files/{id}/download": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "文件下载",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
        "/api/files/{id}/preview": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "文件管理"
                ],
                "summary": "文件预览",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文件ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorRespWrapper"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "admin-server_internal_model.ChunkStatusResp": {
            "type": "object",
            "properties": {
                "uploadedChunks": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "admin-server_internal_model.FileRecord": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "folderId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "mimeType": {
                    "type": "string"
                },
                "originalName": {
                    "type": "string"
                },
                "size": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                },
                "storedName": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "uploaderId": {
                    "type": "integer"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "admin-server_internal_model.MergeFileReq": {
            "type": "object",
            "required": [
                "fileId",
                "fileName",
                "folderId",
                "totalChunks"
            ],
            "properties": {
                "fileId": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "folderId": {
                    "type": "integer"
                },
                "mimeType": {
                    "type": "string"
                },
                "totalChunks": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "admin-server_internal_model.PageResult": {
            "type": "object",
            "properties": {
                "list": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "internal_handler.ChunkStatusRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "00000"
                },
                "data": {
                    "$ref": "#/definitions/admin-server_internal_model.ChunkStatusResp"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "internal_handler.ErrorRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "A0400"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "internal_handler.FileRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "00000"
                },
                "data": {
                    "$ref": "#/definitions/admin-server_internal_model.FileRecord"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "internal_handler.MsgRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "00000"
                },
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                }
            }
        },
        "internal_handler.PageRespWrapper": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "00000"
                },
                "data": {
                    "$ref": "#/definitions/admin-server_internal_model.PageResult"
                },
                "message": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "格式: Bearer {accessToken}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}